Switch integration to DataUpdate Coordinator
